backend/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
apscheduler
holidays
python-dotenv
pydantic[email]
email-validator
